Sum of float numbers

+++ Hello world.rb
@@ -6,4 +6,4 @@ puts "Second number:"
 
 num2 = gets.chomp()
 
-puts (num1.to_i + num2.to_i)
+puts (num1.to_f + num2.to_f)
